Alerts Fix
Better way of doing alerts - not so much duplicated code.
