+++ sciencelabs/alerts/alerts.py
@@ -1,0 +1,22 @@
+#######################################################################################################################
+# Alert stuff helps give user info on changes they make
+
+alert = None  # Default alert to nothing
+
+
+# This method get's the current alert (if there is one) and then resets alert to nothing
+def get_alert():
+    global alert
+    alert_return = alert
+    alert = None
+    return alert_return
+
+
+# This method sets the alert for when one is needed next
+def set_alert(message_type, message):
+    global alert
+    alert = {
+        'type': message_type,
+        'message': message
+    }
+#######################################################################################################################
