+++ sciencelabs/term_startup/__init__.py
@@ -5,29 +5,7 @@ from flask_classy import FlaskView, route
 # Local
 from sciencelabs.term_startup.term_startup_controller import TermStartupController
 from sciencelabs.db_repository.schedule_functions import Schedule
-
-#######################################################################################################################
-# Alert stuff helps give user info on changes they make
-
-alert = None  # Default alert to nothing
-
-
-# This method get's the current alert (if there is one) and then resets alert to nothing
-def get_alert():
-    global alert
-    alert_return = alert
-    alert = None
-    return alert_return
-
-
-# This method sets the alert for when one is needed next
-def set_alert(message_type, message):
-    global alert
-    alert = {
-        'type': message_type,
-        'message': message
-    }
-#######################################################################################################################
+from sciencelabs.alerts.alerts import *
 
 
 class TermStartupView(FlaskView):
